Encourage users to use their default Ruby version
Using the same version of Ruby as GitHub uses isn't essential,
and the pages can be previewed just fine using a different version.

Suggest users try the simple option first, but retain instructions for
installing a different Ruby version (but only for Mac/Linux).

Suggesting that users need a specific version of Ruby, but not giving
instructions on how to achieve this on Windows seems rather inconsistent.